Tor support

+++ build.sh
@@ -1,10 +1,16 @@
 #!/usr/bin/env bash
 set -e
-BITCARTGEN_DOCKER_IMAGE='mrnaif/docker-compose-generator'
-set +e
-docker pull $BITCARTGEN_DOCKER_IMAGE
-docker rmi $(docker images mrnaif/docker-compose-generator --format "{{.Tag}};{{.ID}}" | grep "^<none>" | cut -f2 -d ';') > /dev/null 2>&1
-set -e
+
+: "${BITCARTGEN_DOCKER_IMAGE:=mrnaif/docker-compose-generator}"
+if [ "$BITCARTGEN_DOCKER_IMAGE" == "mrnaif/docker-compose-generator:local" ]
+then
+    docker build generator --tag $BITCARTGEN_DOCKER_IMAGE
+else
+    set +e
+    docker pull $BITCARTGEN_DOCKER_IMAGE
+    docker rmi $(docker images mrnaif/docker-compose-generator --format "{{.Tag}};{{.ID}}" | grep "^<none>" | cut -f2 -d ';') > /dev/null 2>&1
+    set -e
+fi
 
 docker run -v "$PWD/compose:/app/compose" \
     -e "BITCART_INSTALL=${BITCART_INSTALL:-all}" \
